pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "face-recognition"
version = "0.1.0"
description = "Classical face recognition project using PCA and nearest-neighbor methods."
requires-python = ">=3.12"
dependencies = [
  "numpy>=2.0.0,<3.0.0",
  "Pillow>=10.0.0,<12.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "coverage>=7.0.0,<8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["src/tests"]
